refactor: add extension method to register exception middleware
