+++ Src/Fitnet.UnitTests/ExceptionMiddlewareTests.cs
@@ -4,6 +4,7 @@ using System.Net;
 using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 using SuperSimpleArchitecture.Fitnet.Shared.BusinessRulesEngine;
+using ErrorHandling;
 
 public class ExceptionMiddlewareTests
 {
