+++ Src/Fitnet/Program.cs
@@ -1,5 +1,6 @@
 using SuperSimpleArchitecture.Fitnet;
 using SuperSimpleArchitecture.Fitnet.Contracts;
+using SuperSimpleArchitecture.Fitnet.ErrorHandling;
 using SuperSimpleArchitecture.Fitnet.Passes;
 using SuperSimpleArchitecture.Fitnet.Reports;
 using SuperSimpleArchitecture.Fitnet.Shared.Events.EventBus;
@@ -34,7 +35,7 @@ app.UseHttpsRedirection();
 
 app.UseAuthorization();
 
-app.UseMiddleware<ExceptionMiddleware>();
+app.UseErrorHandling();
 
 app.MapControllers();
